// src/store/types.ts
// import NumFactoryStateTypes from './modules/NumFactory/types'

export default interface RootStateTypes {
  isLogin: boolean
  lang: string
  settings: any
  chainInfo: any
  layout: any
  wallet: any
}

// export interface AllStateTypes extends RootStateTypes {
//   numFactoryModule: NumFactoryStateTypes
// }